@@ -1,11 +1,11 @@
 module github.com/odysseia-greek/delphi/periandros
 
-go 1.22
+go 1.23
 
 require (
 	github.com/google/uuid v1.6.0
-	github.com/odysseia-greek/agora/plato v0.1.41
-	github.com/stretchr/testify v1.9.0
+	github.com/odysseia-greek/agora/plato v0.1.48
+	github.com/stretchr/testify v1.10.0
 )
 
 require (
