@@ -4,19 +4,13 @@ import (
 	"github.com/odysseia-greek/agora/diogenes"
 	"github.com/odysseia-greek/agora/plato/config"
 	kubernetes "github.com/odysseia-greek/agora/thales"
+	"os"
 )
 
-func CreateNewConfig(env string) (*PeisistratosHandler, error) {
-	healthCheck := true
-	outOfClusterKube := false
-	debugMode := false
-	if env == "DEVELOPMENT" {
-		healthCheck = false
-		outOfClusterKube = true
-		debugMode = true
-	}
+func CreateNewConfig() (*PeisistratosHandler, error) {
+	env := os.Getenv("ENV")
 
-	vault, err := diogenes.CreateVaultClient(env, healthCheck, debugMode)
+	vault, err := diogenes.CreateVaultClient(true)
 	if err != nil {
 		return nil, err
 	}
@@ -24,7 +18,7 @@ func CreateNewConfig(env string) (*PeisistratosHandler, error) {
 	podName := config.ParsedPodNameFromEnv()
 	ns := config.StringFromEnv(config.EnvNamespace, config.DefaultNamespace)
 
-	kube, err := kubernetes.CreateKubeClient(outOfClusterKube)
+	kube, err := kubernetes.CreateKubeClient(false)
 	if err != nil {
 		return nil, err
 	}
