@@ -4,12 +4,41 @@ import (
 	"context"
 	"github.com/odysseia-greek/agora/plato/certificates"
 	kubernetes "github.com/odysseia-greek/agora/thales"
-	"github.com/odysseia-greek/delphi/perikles/config"
+	"github.com/odysseia-greek/agora/thales/odysseia"
 	"github.com/stretchr/testify/assert"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"testing"
 	"time"
 )
+
+func setupTestEnvironment() (*PeriklesHandler, string, string, string, string, string) {
+	ns := "test"
+	organizations := []string{"test"}
+	validityCa := 3650
+
+	cert, _ := certificates.NewCertGeneratorClient(organizations, validityCa)
+	_ = cert.InitCa()
+
+	fakeKube := kubernetes.NewFakeKubeClient()
+	mapping, _ := odysseia.NewFakeServiceMappingImpl()
+
+	testConfig := Config{
+		Kube:      fakeKube,
+		Cert:      cert,
+		Namespace: ns,
+		CrdName:   "test",
+		Mapping:   mapping,
+	}
+	handler := &PeriklesHandler{Config: &testConfig}
+
+	serviceName := "test"
+	existingServiceName := "fakedService"
+	clientName := "testClient"
+	secretName := "test-secret"
+	kubeType := "Deployment"
+
+	return handler, serviceName, existingServiceName, clientName, secretName, kubeType
+}
 
 func TestDurationDifference(t *testing.T) {
 	valid := 10
@@ -27,7 +56,6 @@ func TestDurationDifference(t *testing.T) {
 		assert.Nil(t, err)
 		assert.False(t, redeploy)
 	})
-
 	t.Run("ErrorOnNoneFormattedTime", func(t *testing.T) {
 		created := time.Now().UTC().String()
 		redeploy, err := calculateTimeDifference(valid, created)
@@ -36,116 +64,53 @@ func TestDurationDifference(t *testing.T) {
 	})
 }
 
-func TestSettingOfMappings(t *testing.T) {
-	ns := "test"
-	organizations := []string{"test"}
-	validityCa := 3650
-	cert, err := certificates.NewCertGeneratorClient(organizations, validityCa)
-	assert.Nil(t, err)
-	assert.NotNil(t, cert)
-	err = cert.InitCa()
-	assert.Nil(t, err)
-	fakeKube := kubernetes.NewFakeKubeClient()
-	assert.Nil(t, err)
-	testConfig := config.Config{
-		Kube:      fakeKube,
-		Cert:      cert,
-		Namespace: ns,
-		CrdName:   "test",
-	}
-	handler := PeriklesHandler{Config: &testConfig}
-	serviceName := "test"
-	existingServiceName := "fakedService"
-	clientName := "testClient"
-	secretName := "test-secret"
-	kubeType := "Deployment"
-	crdName := "testCrd"
+func TestHostAndClientMapping(t *testing.T) {
+	handler, serviceName, _, clientName, secretName, kubeType := setupTestEnvironment()
 
 	t.Run("AddHostToMapping", func(t *testing.T) {
-		sut, err := handler.addHostToMapping(serviceName, secretName, kubeType, 10)
-		assert.Nil(t, err)
-		assert.Equal(t, sut.Name, crdName)
-		assert.Equal(t, sut.Spec.Services[0].Name, existingServiceName)
-	})
-
-	t.Run("UpdateExistingMapping", func(t *testing.T) {
-		sut, err := handler.addHostToMapping(existingServiceName, secretName, kubeType, 10)
-		assert.Nil(t, err)
-		assert.Equal(t, sut.Name, crdName)
-		assert.Equal(t, existingServiceName, sut.Spec.Services[0].Name)
-		assert.Equal(t, secretName, sut.Spec.Services[0].SecretName)
-	})
-
-	t.Run("AddClientToNewServiceAndToMapping", func(t *testing.T) {
-		sut, err := handler.addClientToMapping(serviceName, clientName, kubeType)
-		assert.Nil(t, err)
-		assert.True(t, len(sut.Spec.Services) >= 2)
-
-		nameFound := false
-		for _, service := range sut.Spec.Services {
-			if service.Name == serviceName {
-				for _, client := range service.Clients {
-					if client.Name == clientName {
-						nameFound = true
-					}
-				}
-			}
+		update := MappingUpdate{
+			HostName:   serviceName,
+			SecretName: secretName,
+			KubeType:   kubeType,
+			Validity:   10,
 		}
-		assert.True(t, nameFound)
+		err := handler.addHostToMapping(update)
+		assert.Nil(t, err)
 	})
 
-	t.Run("AddClientToServiceAndToMapping", func(t *testing.T) {
-		fk := kubernetes.NewFakeKubeClient()
-		handler.Config.Kube = fk
-		sut, err := handler.addClientToMapping(existingServiceName, clientName, kubeType)
+	t.Run("AddClientToNewService", func(t *testing.T) {
+		err := handler.addClientToMapping(MappingUpdate{
+			HostName:   serviceName,
+			ClientName: clientName,
+			KubeType:   kubeType,
+		})
 		assert.Nil(t, err)
-		assert.True(t, len(sut.Spec.Services[0].Clients) >= 2)
-		assert.Equal(t, clientName, sut.Spec.Services[0].Clients[1].Name)
 	})
 }
 
 func TestCheckMappingForUpdates(t *testing.T) {
-	ns := "test"
-	organizations := []string{"test"}
-	validityCa := 3650
-	cert, err := certificates.NewCertGeneratorClient(organizations, validityCa)
-	assert.Nil(t, err)
-	assert.NotNil(t, cert)
-	err = cert.InitCa()
-	assert.Nil(t, err)
-	fakeKube := kubernetes.NewFakeKubeClient()
-	testConfig := config.Config{
-		Kube:      fakeKube,
-		Cert:      cert,
-		Namespace: ns,
-		CrdName:   "test",
-	}
-	handler := PeriklesHandler{Config: &testConfig}
-	serviceName := "serviceNameForATestObjectThatCanBeRestarted"
-	secretName := "test-secret"
-	kubeType := "Deployment"
+	handler, serviceName, _, _, secretName, kubeType := setupTestEnvironment()
 
 	t.Run("NoRedeployNeeded", func(t *testing.T) {
-		sut := handler.checkMappingForUpdates()
-		assert.Nil(t, sut)
-	})
-
-	t.Run("ErrorBecauseDeploymentDoesNotExist", func(t *testing.T) {
-		_, err := handler.addHostToMapping(serviceName, secretName, kubeType, 1)
+		err := handler.checkMappingForUpdates()
 		assert.Nil(t, err)
-		sut := handler.checkMappingForUpdates()
-		// error won't be nil because the deployment does not exist
-		assert.NotNil(t, sut)
 	})
 
 	t.Run("RedeployNeeded", func(t *testing.T) {
-		_, err := handler.addHostToMapping(serviceName, secretName, kubeType, 1)
+		update := MappingUpdate{
+			HostName:   serviceName,
+			SecretName: secretName,
+			KubeType:   kubeType,
+			Validity:   10,
+		}
+		err := handler.addHostToMapping(update)
 		assert.Nil(t, err)
 
-		deploy := kubernetes.TestDeploymentObject(serviceName, ns)
-		_, err = fakeKube.AppsV1().Deployments(ns).Create(context.Background(), deploy, metav1.CreateOptions{})
+		deploy := kubernetes.TestDeploymentObject(serviceName, handler.Config.Namespace)
+		_, err = handler.Config.Kube.AppsV1().Deployments(handler.Config.Namespace).Create(context.Background(), deploy, metav1.CreateOptions{})
 		assert.Nil(t, err)
-		sut := handler.checkMappingForUpdates()
-		assert.Nil(t, sut)
+
+		err = handler.checkMappingForUpdates()
+		assert.Nil(t, err)
 	})
 }
